# api/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_noah-isme_sma-adp-api_internal_dto.GenerateScheduleRequest:
    properties:
      classId:
        type: string
      classIds:
        items:
          type: string
        type: array
      days:
        items:
          type: integer
        minItems: 1
        type: array
      hardConstraints:
        items:
          type: string
        type: array
      meta:
        additionalProperties: {}
        type: object
      softConstraints:
        items:
          type: string
        type: array
      subjectLoads:
        items:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_dto.SubjectLoadRequest'
        minItems: 1
        type: array
      termId:
        type: string
      timeSlotsPerDay:
        maximum: 16
        minimum: 1
        type: integer
    required:
    - days
    - subjectLoads
    - termId
    - timeSlotsPerDay
    type: object
  github_com_noah-isme_sma-adp-api_internal_dto.ReportRequest:
    properties:
      classId:
        type: string
      format:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_models.ReportFormat'
      template:
        type: string
      termId:
        type: string
      type:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_models.ReportType'
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_dto.SaveScheduleRequest:
    properties:
      commitToDaily:
        type: boolean
      proposalId:
        type: string
    required:
    - proposalId
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_dto.SubjectLoadRequest:
    properties:
      classId:
        type: string
      difficulty:
        maximum: 10
        minimum: 1
        type: integer
      preferredSlots:
        items:
          type: integer
        type: array
      subjectId:
        type: string
      tags:
        items:
          type: string
        type: array
      teacherId:
        type: string
      weeklyCount:
        minimum: 1
        type: integer
    required:
    - subjectId
    - teacherId
    - weeklyCount
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_models.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_models.ReportFormat:
    enum:
    - csv
    - pdf
    type: string
    x-enum-varnames:
    - ReportFormatCSV
    - ReportFormatPDF
  github_com_noah-isme_sma-adp-api_internal_models.ReportType:
    enum:
    - attendance
    - grades
    - behavior
    - summary
    type: string
    x-enum-varnames:
    - ReportTypeAttendance
    - ReportTypeGrades
    - ReportTypeBehavior
    - ReportTypeSummary
...
  github_com_noah-isme_sma-adp-api_pkg_errors.Error:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  github_com_noah-isme_sma-adp-api_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_errors.Error'
      meta:
        additionalProperties: true
        type: object
      pagination:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_models.Pagination'
    type: object
info:
  contact: {}
  description: Bootstrap server for Golang migration (Phase 0)
  title: SMA ADP API
  version: 0.1.0
paths:
...
  /export/{token}:
    get:
      parameters:
      - description: Signed token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Download generated report via signed token
      tags:
      - Reports
...
  /reports/classes/{id}:
    get:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      - description: Subject ID
        in: query
        name: subjectId
        required: true
        type: string
      - description: Term ID
        in: query
        name: termId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Class grade report
      tags:
      - Reports
  /reports/generate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Report request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_dto.ReportRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Queue a new report job
      tags:
      - Reports
  /reports/status/{id}:
    get:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get report job status
      tags:
      - Reports
  /reports/students/{id}:
    get:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      - description: Term ID
        in: query
        name: termId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Student report card
      tags:
      - Reports
  /schedule/generate:
    post:
      consumes:
      - application/json
      description: Legacy path kept for backward compatibility. Prefer /schedules/generator
        for new integrations.
      parameters:
      - description: Generate schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_dto.GenerateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Generate conflict-free schedule proposal (legacy endpoint)
      tags:
      - Academics
  /schedule/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: Save schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_dto.SaveScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Save schedule proposal to semester schedules
      tags:
      - Scheduler
...
  /schedules/generator:
    post:
      consumes:
      - application/json
      description: Preferred endpoint for UI preview mode. Responses include mode
        metadata to distinguish preview vs. persisted schedules.
      parameters:
      - description: Generate schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_dto.GenerateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Generate schedule proposal (canonical alias)
      tags:
      - Academics
...
  /semester-schedule:
    get:
      parameters:
      - description: Term ID
        in: query
        name: termId
        required: true
        type: string
      - description: Class ID
        in: query
        name: classId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List semester schedules for class-term
      tags:
      - Scheduler
  /semester-schedule/{id}:
    delete:
      parameters:
      - description: Semester schedule ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete draft semester schedule
      tags:
      - Scheduler
  /semester-schedule/{id}/slots:
    get:
      parameters:
      - description: Semester schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get slots for a semester schedule
      tags:
      - Scheduler
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"